+++ src/test/java/br/com/fiap/restaurante/exception/GlobalExceptionHandlerTest.java
@@ -5,6 +5,7 @@ import br.com.fiap.restaurante.model.Address;
 import br.com.fiap.restaurante.model.User;
 import br.com.fiap.restaurante.model.UserType;
 import br.com.fiap.restaurante.security.JwtService;
+import br.com.fiap.restaurante.service.TipoUsuarioService;
 import br.com.fiap.restaurante.service.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,6 +38,9 @@ class GlobalExceptionHandlerTest {
 
     @MockBean
     private JwtService jwtService;
+
+    @MockBean
+    private TipoUsuarioService tipoUsuarioService;
 
     private User user;